# packages/qwksearch-api-client/qwksearch-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QwkSearch Research Agent API",
    "version": "1.0.0",
    "description": "API for QwkSearch — AI-powered research with search, document management, and multi-provider LLM support."
  },
  "servers": [
    { "url": "https://qwksearch.com/api", "description": "Production" }
  ],
...
  "paths": {
...
    "/doc/documents/{id}": {
      "get": {
        "tags": ["Documents"],
        "summary": "Get a document by ID",
        "operationId": "getDocumentById",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Document",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Document" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Documents"],
        "summary": "Update a document",
        "operationId": "updateDocument",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "parentId": { "type": "integer" },
                  "isExpanded": { "type": "boolean" },
                  "metadata": { "type": "object" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated document",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Document" } } }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Documents"],
        "summary": "Delete a document",
        "operationId": "deleteDocument",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" }
        }
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "name": { "type": "string" },
          "content": { "type": "string" },
          "parentId": { "type": "integer" },
          "isFolder": { "type": "boolean" },
          "isExpanded": { "type": "boolean" },
          "metadata": { "type": "object" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "userId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}